fix remaining time and profile changing signals, to match those in PowerDevil (battery applet fixes coming up next)
svn path=/trunk/KDE/kdebase/workspace/; revision=1201015

+++ plasma/generic/dataengines/powermanagement/powermanagementengine.h
@@ -54,7 +54,7 @@ private slots:
     void deviceRemoved(const QString& udi);
     void deviceAdded(const QString& udi);
     void profileChanged(const QString &current);
-    void batteryRemainingTimeChanged(int time);
+    void batteryRemainingTimeChanged(qulonglong time);
     void availableProfilesChanged();
     void reloadPowerDevilData();
